requirements.txt
coverage==7.1.0
pipenv==2022.12.19
pytest==7.2.1
Faker==16.7.0
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
marshmallow-sqlalchemy==0.28.1
SQLAlchemy==1.4.42
gunicorn==20.1.0
psycopg2-binary==2.9.4
marshmallow~=3.20.1
requests==2.28.2
boto3==1.20.17